Fix movie playback bug.
Previously there was a bug where if you save a keyframe with a button pressed and then seek back to re-record the follow-up without the button pressed, then the button would end up pressed anyway when playing back through the keyframe.  This could be mitigated by always restoring the snapshot, but I'd rather not resort to that.  Instead, we fix the underlying issue by resetting the controller when playing back through a keyframe, since any "already-pressed" buttons will be re-emitted at the start of the chunk, anyway.