Massive performance improvements.
At this point, with debugging disabled and frameskip set to 60, I can average just under 800 fps on my macbook, which is pretty impressive.  The biggest single resource hog outside CPU.emulate is PAPU.clockFrameCounter, which does a lot of adding and subtracting but is necessary to correctly emulate sound IRQs and status bits, since we need to know when the samples happen to clock the length counter correctly.